1025 ADD exercise board 카카오지도api 기능 구현
address.jsp를 write.jsp include해서 가져온 다음 위도와 경도 변수를 선언하여 document.getElementById("lat").value = lat; 하여 인풋값에 넣을 수 있게 하였다.

+++ src/main/java/com/hong/fitness/vo/FitnessBoardVO.java
@@ -18,6 +18,6 @@ public class FitnessBoardVO {
 	private int reply_count;
 	
 	// 위도, 경도 컬럼 추가
-	private BigDecimal lat;
-	private BigDecimal lng;
+	private Double lat;
+	private Double lng;
 }
